+++ example/lib/main.dart
@@ -21,7 +21,7 @@ class MyApp extends StatelessWidget {
         useMaterial3: true,
       ),
       darkTheme: ThemeData.dark(),
-      home: JapanStyle(),
+      home: InheritsParentBuilder(),
       // home: const Music(
       //   cover:
       //       "https://static.wikia.nocookie.net/iowiro/images/c/c5/News_10202019a.jpg",